Add getArgs and support pull option in testE2e
