+++ .eslintrc
@@ -16,6 +16,7 @@
           "**/packages/psss/src/api/FakeApi.js",
           "**/packages/psss/src/utils/test-utils.js",
           "**/packages/ui-components/helpers/testingRenderer.js",
+          "**/packages/utils/src/script.js",
           "**/packages/web-frontend/config/**",
           "**/packages/web-frontend/scripts/**"
         ],
