+++ scripts/node/Script.js
@@ -6,9 +6,8 @@
 const { execSync } = require('child_process');
 const { existsSync, readdirSync } = require('fs');
 const { platform } = require('os');
-const yargs = require('yargs');
 
-const { getLoggerInstance } = require('@tupaia/utils');
+const { getArgs, getLoggerInstance } = require('@tupaia/utils');
 
 /**
  * @abstract
@@ -21,19 +20,6 @@ class Script {
 
   logger;
 
-  /**
-   * @protected Override in child class
-   *
-   * @property {string} [command] Use '*' or '$0' to refer to the default command
-   *   This field is useful for
-   *   1. Defining subcommands, eg 'push <branch_name>'
-   *   2. Defining positional arguments in the default command, eg '* <path>'
-   * @property {Object} [options]
-   * @property {string} [usage]
-   * @property {string} [version]
-   *
-   * @see https://github.com/yargs/yargs/blob/master/docs/api.md
-   */
   config = {};
 
   args;
@@ -48,15 +34,7 @@ class Script {
   }
 
   parseConfig() {
-    const supportedConfigKeys = ['command', 'options', 'usage', 'version'];
-
-    yargs.strict();
-    Object.entries(this.config).forEach(([key, value]) => {
-      if (supportedConfigKeys.includes(key)) {
-        yargs[key](value);
-      }
-    });
-    this.args = yargs.argv;
+    this.args = getArgs(this.config);
   }
 
   /**
